\name{geckoCreateAccount}
\alias{geckoCreateAccount}

\title{Create Geckoboard Account}

\description{Create object holding meta data of a Geckoboard project.}

\usage{geckoCreateAccount(name, key)}

\arguments{
  \item{name}{Can be arbitrary.}
  \item{key}{API key of Geckoboard project.}
}

\value{An object holding all the values specified as parameters.}

\author{Meinhard Ploner}

\seealso{
\code{\link{geckoDatasetCreate}}
}

\keyword{geckoboard}
\keyword{account}
