\name{geckoImageSave}
\alias{geckoImageSave}

\title{Helper function to get an image created for Geckoboard Image}

\description{Helper function to get an image created for Geckoboard Image. Call this *after* creating a plot using e.g. \code{plot} or \code{ggplot}.}

\usage{geckoImageSave()}

\value{Nothing is returned.}

\author{Meinhard Ploner}

\keyword{geckoboard}
\keyword{image}
