\name{RGeckoboard-package}
\alias{RGeckoboard-package}
\alias{RGeckoboard}
\docType{package}

\title{\packageTitle{RGeckoboard}}

\description{\packageDescription{RGeckoboard}}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{RGeckoboard}
\packageIndices{RGeckoboard}
The package provides easy access to the Geckoboard API from R.
}

\author{
\packageAuthor{RGeckoboard}
Maintainer: \packageMaintainer{RGeckoboard}
}

\references{\url{https://www.geckoboard.com}}

\keyword{package}
\keyword{Geckoboard}
\keyword{API}
\keyword{dashboard}
