\name{geckoDatasetRemove}
\alias{geckoDatasetRemove}

\title{Remove a Dataset from Geckoboard}

\description{Remove a Dataset from the Geckoboard server.}

\usage{geckoDatasetRemove(apiKey, dataset="dummy.by_day")}

\arguments{
  \item{apiKey}{API key.}
  \item{dataset}{Name of an existing Geckoboard dataset.}
}
  
\value{Nothing is returned.}

\author{Meinhard Ploner}

\keyword{geckoboard}
\keyword{dataset}
